Refactor, move generators to data_generators. Investigating image discontinuities
